Правки на основе ручного тестирования

+++ Aworkplace/Views/listLiterature.cs
@@ -15,13 +15,15 @@ namespace Aworkplace.Views
 {
     public partial class listLiterature : Form
     {
-        List<TypeLiterature> allLiteratures = new List<TypeLiterature>();
-        Dictionary<Int32, String> typeLiterature = new Dictionary<Int32, String>();
+        private List<TypeLiterature> allLiteratures = new List<TypeLiterature>();
+        private Dictionary<Int32, String> typeLiterature = new Dictionary<Int32, String>();
         private readonly Functions f = new Functions();
 
         public listLiterature()
         {
-            InitializeComponent();            
+            InitializeComponent();
+            this.MaximizeBox = false;
+            this.FormBorderStyle = FormBorderStyle.FixedSingle;
             f.readFromFileForData(ref dataLiterature, ref allLiteratures, ref typeLiterature);
         }
 
@@ -32,15 +34,20 @@ namespace Aworkplace.Views
 
         private void deleteSelectedItemButton_Click(object sender, EventArgs e)
         {
-            if (dataLiterature.SelectedCells[0].RowIndex != -1)
+            var result = MessageBox.Show("Вы точно хотите удалить читателя?", "Внимание",
+            MessageBoxButtons.YesNo, MessageBoxIcon.Question);
+            if (result == DialogResult.Yes)
             {
-                allLiteratures[dataLiterature.SelectedCells[0].RowIndex].DeleteLiterature();
-                f.readFromFileForData(ref dataLiterature, ref allLiteratures, ref typeLiterature);
-            }
-            else
-            {
-                MessageBox.Show("Выделите одну строку в таблице!");
-            }
+                if (dataLiterature.SelectedCells[0].RowIndex != -1)
+                {
+                    allLiteratures[dataLiterature.SelectedCells[0].RowIndex].DeleteLiterature();
+                    f.readFromFileForData(ref dataLiterature, ref allLiteratures, ref typeLiterature);
+                }
+                else
+                {
+                    MessageBox.Show("Выделите одну строку в таблице!");
+                }
+            }                
         }
 
         private async void addLiteratureButton_Click(object sender, EventArgs e)
